% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_algos.R
\name{all_algos}
\alias{all_algos}
\title{Run all measurement cleaning algorithms simultaneously}
\usage{
all_algos(
  df,
  id,
  measures,
  tmeasures,
  start_point = NULL,
  variables = NULL,
  basic = FALSE
)
}
\arguments{
\item{df}{\code{data.frame} holding raw (input) measurements.}

\item{id}{string corresponding to the name of the column of group identifiers
in \code{df}.}

\item{measures}{string corresponding to the name of the column of measures in
\code{df}, e.g., numeric weight data if using to clean weight data.}

\item{tmeasures}{string corresponding to the name of the column of measure
dates and/or times in \code{df}.}

\item{start_point}{string corresponding to the name of the column in \code{df}
holding the time at which subsequent measurement dates will be assessed,
should be the same for each group. Eg., if \code{t = 0} (\code{t[1]})
corresponds to an index visit held by the variable \code{VisitDate}, then
\code{startPoint} should be set to \code{VisitDate}.}

\item{variables}{Optional. Specific to the Maguen *et al.* 2013 algorithm,
character vector describing the terms in \code{df} to include on the RHS
of the internal mixed model. E.g., \code{c("Age", "Gender")} would
generate a model of the form \code{measures ~ Age + Gender + (1|id)}.}

\item{basic}{run only those algorithms involving **just** \code{id},
\code{measures}, \code{tmeasures}.}
}
\value{
named list with processed data for each algorithm.
}
\description{
\code{all_algos} runs all 12 algorithms with default settings and outputs a
named list with each result. Each algorithm requires a basic set of data:
a patient identifier, \code{id}; the measures to clean \code{measures}
(e.g., weight, continuous lab data, etc.); a recorded time at which the
measure was taken \code{tmeasures}. Some require an index date,
\code{StartPoint} (Janney *et al.* 2016, Goodrich *et al.* 2016, Jackson
*et al.* 2015, Kazerooni & Lim 2016, Rosenberger *et al.* 2011), Maguen
*et al.* 2013 adjusts for optional demographic variables (e.g., age, sex).
Buta *et al.* 2018 stands out, requiring BMI measurements as the unit of
measurement \code{measures}. Thus, in order to fully utilize this function,
\code{df} needs to have at least \code{id}, \code{measures},
\code{tmeasures}, \code{StartPoint}, and BMI.
}
