% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maciejewski2016.R
\name{consolidate_sameday_records}
\alias{consolidate_sameday_records}
\title{Maciejewski algorithm part I.b. - consolidate same day records}
\usage{
consolidate_sameday_records(
  df,
  id,
  tmeasures,
  measures,
  varvec = NULL,
  keep_uncleanable = TRUE,
  outliers = list(LB = 75L, UB = 700L),
  sd_thresholds = list(same_day = 2)
)
}
\arguments{
\item{df}{object of class \code{data.frame}, containing \code{id} and
\code{measures}}

\item{id}{string corresponding to the name of the column of patient
identifiers in \code{df}.}

\item{tmeasures}{string corresponding to the name of the column of measure
dates and/or times in \code{df}}

\item{measures}{string corresponding to the name of the column of measures in
\code{df}, e.g., numeric weight data if using to clean weight data.}

\item{varvec}{optional. String vector containing \code{df}, \code{id},
\code{measures}, and \code{tmeasures}.}

\item{keep_uncleanable}{subjects having exactly two \code{tmeasures}
(internally `measurement_date`) values each with a standard deviation of the
\code{measures} (internally `measurement`) values that is strictly greater
than 2, are termed "uncleanable", the user can choose to keep or remove the
"uncleanable" subjects. The default is to keep them/\code{TRUE}.}

\item{outliers}{object of type \code{list} with two values, \code{UB} and
\code{LB}, the lower and upper bounds of plausible values of
\code{measures}. Default is \code{list(`LB` = 75L, `UB` = 700L)}.}

\item{sd_thresholds}{Object of type \code{list} with 1 numeric value of
same-day standard deviation statistics above which the data of the
associated \code{id}-\code{tmeasures} combination are considered indicative
of a discrepancy in \code{measures} values that requires more scrutiny to
resolve into one unique \code{measures} value per
\code{id}-\code{tmeasures} combination than simply taking the arithmetic
mean of the values.}
}
\value{
\code{data.table} containing the three variables supplied to the
  input parameters (\code{id}, \code{tmeasures}, \code{measures}). The data
  set contains one value of \code{measures} per \code{id} and
  \code{tmeasures} combination. There is also the option to remove
  intermediate data sets created to produce the main output data set.
}
\description{
\code{consolidate_sameday_records} reduces input data set to one unique
\code{measures} value per combination of \code{id} and \code{tmeasures}.
}
\examples{
\dontrun{
# Checking performance
data(cdw1000)

# with defaults
consolidate_sameday_records(
  df = cdw1000,
  id = 'id',
  tmeasures = 'WeightDate',
  measures = 'Weight'
)

# removing "uncleanable" records
consolidate_sameday_records(
  df = cdw1000,
  id = 'id',
  tmeasures = 'WeightDate',
  measures = 'Weight',
  keep_uncleanable = FALSE
)

# with varvec
varvec <- c(
  'id' = 'id',
  'tmeasures' = 'WeightDate',
  'measures' = 'Weight'
)
consolidate_sameday_records(
  df = cdw1000,
  varvec = varvec
)
}
}
