% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Janney2016.R
\name{janney}
\alias{janney}
\title{Janney et al. 2016 Weight Measurment Cleaning Algorithm}
\usage{
janney(
  df,
  id,
  measures,
  tmeasures,
  startPoint,
  t = c(0, 182.5, 365),
  windows = c(30, 60, 60),
  outliers = list(LB = c(91, 72, 72), UB = c(600, 650, 650))
)
}
\arguments{
\item{df}{object of class \code{data.frame}, containing \code{id} and
\code{weights}.}

\item{id}{string corresponding to the name of the column of group identifiers
in \code{df}.}

\item{measures}{string corresponding to the name of the column of measures in
\code{df}, e.g., numeric weight data if using to clean weight data.}

\item{tmeasures}{string corresponding to the name of the column of measure
dates and/or times in \code{df}.}

\item{startPoint}{string corresponding to the name of the column in \code{df}
holding the time at which subsequent measurement dates will be assessed,
should be the same for each group. Eg., if \code{t = 0} (\code{t[1]})
corresponds to an index visit held by the variable \code{VisitDate}, then
\code{startPoint} should be set to \code{VisitDate}.}

\item{t}{numeric vector of time points to collect measurements, eg.
\code{c(0, 182.5, 365)} for measure collection at \code{t = 0},
\code{t = 180} (6 months from \code{t = 0}), and \code{t = 365} (1 year
from \code{t = 0}). Default is \code{c(0, 182.5, 365)} according to
Janney et al. 2016}

\item{windows}{numeric vector of measurement collection windows to use around
each time point in \code{t}. Eg. Janney et al. 2016 use
\code{c(30, 60, 60)} for \code{t} of \code{c(0, 182.5, 365)}, implying that
the closest measurement \code{t = 0} will be collected 30 days prior to and
30 days post startPoint. Subsequent measurements will be collected 60 days
prior to and 60 days post t0+182.5 days, and t0+365 days.}

\item{outliers}{optional. object of type \code{list} with numeric inputs
corresponding to the upper and lower bound for each time entry in parameter
\code{t}. Default is \code{list(LB = c(91, 72, 72), UB = c(600, 650, 650))}
for \code{t = c(0, 182.56, 365)}, differing between baseline and subsequent
measurment collection dates. If not specified then only the subsetting and
window functions will be applied.}
}
\value{
returns a data frame reduced by the inclusion criteria defined in
  \code{t}, \code{windows} and \code{outliers}, the resultant dimensions
  being \code{ncol(df) + 2} columns and up to \code{length(t)} rows per
  group. Output measurements held in column \code{measout}.
}
\description{
For grouped time series, (e.g., per person) Define time points and collect
measurements, optionally applying windows for possible data points for each
time point, then removing outliers. Based on work by Janney et al. 2016.
}
\examples{
library(dplyr)
data(cdw1000)

janney_df <- janney(
   df = cdw1000,
   id = "id",
   measures = "Weight",
   tmeasures = "WeightDate",
   startPoint = "VisitDate"
  )

# dplyr::glimpse(janney_df)

compare_df <- bind_rows(
  "Input"  = cdw1000 \%>\% select(Weight),
  "Output" = janney_df \%>\%
     select(measout) \%>\%
     na.omit() \%>\%
     rename(Weight = measout),
  .id = "IO"
)

compare_df \%>\%
  group_by(IO) \%>\%
  summarize(
    n      = n(),
    mean   = mean(Weight),
    sd     = sd(Weight),
    min    = min(Weight),
    median = median(Weight),
    max    = max(Weight)
  )
}
