% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Littman2012.R
\name{littman}
\alias{littman}
\title{Littman et al. 2012 Measurment Cleaning Algorithm}
\usage{
littman(
  df,
  id,
  measures,
  tmeasures,
  outliers = list(LB = c(75), UB = c(600)),
  sd_threshold = 0.1,
  add_internals = FALSE
)
}
\arguments{
\item{df}{object of class \code{data.frame}, containing \code{id} and
\code{measures}.}

\item{id}{string corresponding to the name of the column of patient
identifiers in \code{df}.}

\item{measures}{string corresponding to the name of the column of
measurements in \code{df}.}

\item{tmeasures}{string corresponding to the name of the column of
measurement collection dates or times in \code{df}. If \code{tmeasures} is
a date object, there may be more than one measure on the same day, if it is
a precise datetime object, there may not be more than one measure on the
same day.}

\item{outliers}{object of type \code{list} with numeric inputs corresponding
to the upper and lower bound for each time entry. Default is
\code{list(LB = c(75), UB = c(600))}.}

\item{sd_threshold}{numeric scalar to be multiplied by the \code{meanMeasures}
per \code{id}. E.g., from Littman 2012, "...We excluded any weight
measurements that met the following 2 criteria: 1) the difference between
the mean weight and weight in question was greater than the SD and 2) the
SD was greater than 10 percent of the mean...." implies
\code{sd_threshold = 0.1}.}

\item{add_internals}{logical, adds additional columns to output data frame
detailing the group-wise means and standard deviations used interally to
process the data. Defaults to FALSE.}
}
\value{
input data.frame with additional column, \code{measout}, the
  output/"cleaned" data. If \code{add_internals == TRUE}, 5 additional columns
  are added to the resultant data frame: code{meanMeasures},
  \code{mean(measures)} per \code{id}; \code{SDMeasures}, \code{SD(measures)}
  per \code{id}; \code{SD_threshold_}, \code{mean(measures) * sd_threshold};
  and \code{cond1} & \code{cond2}, logical vectors indicating the result of
  checking the two conditions; 1, the difference between the
  \code{meanMeasures} and code{measures_aug_} in question is greater than the
  \code{SDMeasures}, and 2, the \code{SDMeasures} is greater than
  \code{SD_threshold_} of the mean.
}
\description{
For grouped time series, (e.g., per person) clean numeric measurements
collected longitudinally, first, removing outliers, then examining within-
person measurement mean and standard deviation, if measurement is found to be
above \code{sd_threshold}, then measurement is set to missing (\code{NA}).
Based on work by Littman et al. 2012.
}
\examples{
library(dplyr)
library(ggplot2)
data(cdw1000)

littman_df <- littman(
   df = cdw1000,
   id = "id",
   measures = "Weight",
   tmeasures = "WeightDate"
  )

# dplyr::glimpse(littman_df)

samp <- littman_df \%>\%
  distinct(id) \%>\%
  sample_n(16) \%>\%
  pull()

littman_df \%>\%
  filter(id \%in\% samp) \%>\%
  ggplot(aes(x = WeightDate, y = measout)) +
  geom_point(color = "green") +
  geom_line() +
  facet_wrap(vars(id), ncol = 4)
}
