% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maciejewski2016.R
\name{maciejewski}
\alias{maciejewski}
\title{Maciejewski algorithm complete}
\usage{
maciejewski(
  df,
  id,
  measures,
  tmeasures,
  varvec = NULL,
  n_cleaning_iters = 1L,
  keep_uncleanable = TRUE,
  outliers = list(LB = 75L, UB = 700L),
  sd_thresholds = list(same_day = 2L, high_sd_flag = 20L, first_last = 35L, abc_low =
    10L, abc_high = 45L),
  skip_consolidation = FALSE,
  record_execution_times = FALSE,
  print_execution_times = record_execution_times
)
}
\arguments{
\item{df}{object of class \code{data.table}, containing \code{id} and
\code{measures}}

\item{id}{string corresponding to the name of the column of patient
identifiers in \code{df}.}

\item{measures}{string corresponding to the name of the column of measures in
\code{df}, e.g., numeric weight data if using to clean weight data.}

\item{tmeasures}{string corresponding to the name of the column of measure
dates and/or times in \code{df}}

\item{varvec}{optional. String vector containing \code{df}, \code{id},
\code{measures}, and \code{tmeasures}.}

\item{n_cleaning_iters}{number of times to pass data through the cleaning
algorithm.}

\item{keep_uncleanable}{subjects having exactly two \code{tmeasures}
(internally `measurement_date`) values each with a standard deviation of the
\code{measures} (internally `measurement`) values that is strictly greater
than 2, are termed "uncleanable", the user can choose to keep or remove the
"uncleanable" subjects. The default is to keep them/\code{TRUE}.}

\item{outliers}{object of type \code{list} with two values, \code{UB} and
\code{LB}, the lower and upper bounds of plausible values of
\code{measures}. Default is \code{list(`LB` = 75L, `UB` = 700L)}.}

\item{sd_thresholds}{Object of type \code{list} with 5 values
\itemize{
   \item same_day - numeric value of same-day standard deviation statistics
   above which the data of the associated \code{id}-\code{tmeasures}
   combination are considered indicative of a discrepancy in \code{measures}
   values that requires more scrutiny to resolve into one unique
   \code{measures} value per \code{id}-\code{tmeasures} combination than
   simply taking the arithmetic mean of the values.
   \item high_sd_flag - Numeric value of the rolling standard deviation
   statistic above which the \code{measures} values of the three associated
   \code{id}-\code{tmeasures} combinations are considered indicative of
   containing one or more "outlier" values.
   \item first_last - Numeric value of the rolling standard deviation
   statistic above which the \code{measures} values of the three associated
   \code{id}-\code{tmeasures} combinations are considered indicative of
   containing one or more "outlier" values. This value is used only to detect
   outliers among the first three and the last three
   \code{id}-\code{tmeasures} combinations, in chronologically ascending
   order.
   \item abc_low - Numeric value of the pairwise standard deviation statistic
   below which the \code{measures} values of the two associated
   \code{id}-\code{tmeasures} combinations are considered not indicative of
   containing an "outlier" value. This value is used only to detect outliers
   in all observations except the first three or the last three
   \code{id}-\code{tmeasures} combinations, in chronologically ascending
   order. This argument is used in conjunction with the \code{abc_high}
   argument.
   \item abc_high - Numeric value of the pairwise standard deviation
   statistic above which the \code{measures} values of the two associated
   \code{id}-\code{tmeasures} combinations are considered indicative of
   containing an "outlier" value. This value is used only to detect outliers
   in all observations except the first three or the last three
   \code{id}-\code{tmeasures} combinations, in chronologically ascending
   order. This argument is used in conjunction with the \code{abc_low}
   argument.
}}

\item{skip_consolidation}{bypass consolidation step for datasets that
already have one record per date per person. Default is \code{FALSE}.}

\item{record_execution_times}{record computation time? Default is
\code{FALSE}.}

\item{print_execution_times}{print the recorded computation time? Default is
set to the value of \code{record_execution_times}.}
}
\value{
\code{data.table} with measurements cleaned according to Maciejewski
  et al. 2016.
}
\description{
\code{maciejewski} applies both part I and II of the Maciejewski 2016
algorithm to a "raw" input dataset, it is basically \code{maciejewski_algo}
with additional cleaning processes. The various threshold values used
throughout these macros were selected via an iterative process where they
were (subjectively) judged to provide the best balance between removing
highly improbable weight measurement values and discounting true drastic
changes in weight over time. If you believe that these threshold values
might not be optimal for your data, please feel free to adjust them to suit
your needs; they need not be considered the absolute "gold standard",
especially if better criteria are found.
}
\examples{
\dontrun{
data(cdw1000)

maciejewski.dt <- maciejewski(
  df = cdw1000,
  id = 'id',
  tmeasures = 'WeightDate',
  measures = 'Weight'
)

glimpse(maciejewski.dt)

maciejewski.dt <- maciejewski(
  df = cdw1000,
  id = 'id',
  tmeasures = 'WeightDate',
  measures = 'Weight',
  record_execution_times = TRUE
)

maciejewski.dt <- maciejewski(
  df = cdw1000,
  id = 'id',
  tmeasures = 'WeightDate',
  measures = 'Weight',
  skip_consolidation = TRUE,
  record_execution_times = TRUE
)
}
}
