% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maciejewski2016.R
\name{maciejewski_algo}
\alias{maciejewski_algo}
\title{Maciejewski algorithm part II. - "clean" records}
\usage{
maciejewski_algo(
  df,
  id,
  tmeasures,
  measures,
  varvec = NULL,
  sd_thresholds = list(same_day = 2L, high_sd_flag = 20L, first_last = 35L, abc_low =
    10L, abc_high = 45L)
)
}
\arguments{
\item{df}{object of class \code{data.table}, containing \code{id} and
\code{measures}}

\item{id}{string corresponding to the name of the column of patient
identifiers in \code{df}.}

\item{tmeasures}{string corresponding to the name of the column of measure
dates and/or times in \code{df}}

\item{measures}{string corresponding to the name of the column of measures in
\code{df}, e.g., numeric weight data if using to clean weight data.}

\item{varvec}{optional. String vector containing \code{df}, \code{id},
\code{measures}, and \code{tmeasures}.}

\item{sd_thresholds}{Object of type \code{list} with 5 values
\itemize{
   \item same_day - numeric value of same-day standard deviation statistics
   above which the data of the associated \code{id}-\code{tmeasures}
   combination are considered indicative of a discrepancy in \code{measures}
   values that requires more scrutiny to resolve into one unique
   \code{measures} value per \code{id}-\code{tmeasures} combination than
   simply taking the arithmetic mean of the values.
   \item high_sd_flag - Numeric value of the rolling standard deviation
   statistic above which the \code{measures} values of the three associated
   \code{id}-\code{tmeasures} combinations are considered indicative of
   containing one or more "outlier" values.
   \item first_last - Numeric value of the rolling standard deviation
   statistic above which the \code{measures} values of the three associated
   \code{id}-\code{tmeasures} combinations are considered indicative of
   containing one or more "outlier" values. This value is used only to detect
   outliers among the first three and the last three
   \code{id}-\code{tmeasures} combinations, in chronologically ascending
   order.
   \item abc_low - Numeric value of the pairwise standard deviation statistic
   below which the \code{measures} values of the two associated
   \code{id}-\code{tmeasures} combinations are considered not indicative of
   containing an "outlier" value. This value is used only to detect outliers
   in all observations except the first three or the last three
   \code{id}-\code{tmeasures} combinations, in chronologically ascending
   order. This argument is used in conjunction with the \code{abc_high}
   argument.
   \item abc_high - Numeric value of the pairwise standard deviation
   statistic above which the \code{measures} values of the two associated
   \code{id}-\code{tmeasures} combinations are considered indicative of
   containing an "outlier" value. This value is used only to detect outliers
   in all observations except the first three or the last three
   \code{id}-\code{tmeasures} combinations, in chronologically ascending
   order. This argument is used in conjunction with the \code{abc_low}
   argument.
}}
}
\value{
\code{data.table} with measurements cleaned according to Maciejewski
  et al. 2016.
}
\description{
\code{maciejewski_algo} applies "cleaning" algorithm to \code{measures}
values, removing suspected outliers.
}
\examples{
\dontrun{
data(cdw1000)

# with defaults
consolidated.dt <- consolidate_sameday_records(
  df = cdw1000,
  id = 'id',
  tmeasures = 'WeightDate',
  measures = 'Weight'
)

maciejewski.dt <- maciejewski_algo(
  df = consolidated.dt,
  id = 'id',
  tmeasures = 'WeightDate',
  measures = 'Weight'
)

glimpse(maciejewski.dt)

uniqueN(maciejewski.dt[, .(id, WeightDate)])
}
}
