% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maguen2013.R
\name{maguen}
\alias{maguen}
\title{Maguen 2013 Measurment Cleaning Algorithm}
\usage{
maguen(
  df,
  id,
  measures,
  tmeasures,
  outliers = list(LB = 70, UB = 700),
  variables = NULL,
  resid_threshold = 10,
  add_internals = FALSE,
  ...
)
}
\arguments{
\item{df}{object of class \code{data.frame}, containing \code{id} and
\code{measures}.}

\item{id}{string corresponding to the name of the column of group identifiers
in \code{df}.}

\item{measures}{string corresponding to the name of the column of
measurements in \code{df}.}

\item{tmeasures}{string corresponding to the name of the column of
measurement collection dates or times in \code{df}. If \code{tmeasures} is
a date object, there may be more than one weight on the same day, if it is
a precise \code{datetime} object, there may not be more than one weight on
the same day.}

\item{outliers}{object of type \code{list} with numeric inputs corresponding
to the upper and lower bound for each time entry. Default is
\code{list(LB = c(70), UB = c(700))}.}

\item{variables}{character vector describing the terms in \code{df} to
include on the RHS of the internal mixed model. E.g.,
\code{c("Age", "Gender")} would generate a model of the form
\code{measures ~ Age + Gender + (1|id)}.}

\item{resid_threshold}{single numeric value to be used as a cut off value for
the conditional (response) residual for each measurement. Lower values are
more conservative.}

\item{add_internals}{logical, adds additional columns to output data frame:
\code{Output}, input data with outliers set to \code{NA}; \code{t}, time
standardized for each group in \code{id}; \code{resid}, the residual from
the result of \code{lme4::lmer}. Defaults to FALSE.}

\item{...}{any number of named arguments for the \code{lme4::lmer} model}
}
\value{
if add_internals is FALSE, returns input data frame with processed
  data as an additional column \code{measout}. if \code{add_internals} is TRUE
  then it returns the input data frame with processed data \code{measout} and
  columns described in \code{add_internals}.
}
\description{
For grouped time series, (e.g., per person) \code{maguen} removes outliers,
then models patients weight trajectories with a linear mixed model with a
random slope and intercept, adjusted for age and gender, it then removes
measurements where the residual meets a certain threshold. Based on work by
Maguen et al. 2013
}
\examples{
library(dplyr)
library(ggplot2)
data(cdw1000)

maguen_df <- maguen(
   df = cdw1000,
   id = "id",
   measures = "Weight",
   tmeasures = "WeightDate",
   variables = c("Age", "Sex"),
   add_internals = TRUE
  )

# dplyr::glimpse(maguen_df)

samp <- maguen_df \%>\%
  filter(is.na(measout)) \%>\%
  distinct(id) \%>\%
  sample_n(16) \%>\%
  pull()

maguen_df \%>\%
  filter(id \%in\% samp) \%>\%
  ggplot() +
  geom_line(aes(x = t, y = Weight), color = "black") +
  geom_point(aes(x = t, y = measout), color = "green") +
  facet_wrap(vars(id), ncol = 4, scales = "free")
}
