% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Janney2016.R
\name{meas_collect}
\alias{meas_collect}
\title{Collect Weight Measures from Arbitrary Time-Points}
\usage{
meas_collect(
  df,
  id,
  measures,
  tmeasures,
  startPoint,
  t = c(0, 182, 365),
  windows = c(30, 60, 60)
)
}
\arguments{
\item{df}{object of class \code{data.frame}, containing \code{id} and
\code{measures}.}

\item{id}{string corresponding to the name of the column of patient
identifiers in \code{df}.}

\item{measures}{string corresponding to the name of the column of measures
in \code{df}, e.g., numeric weight data if using to clean weight data.}

\item{tmeasures}{string corresponding to the name of the column of measure
dates and/or times in \code{df}.}

\item{startPoint}{string corresponding to the name of the column in \code{df}
holding the time at which subsequent measurement dates will be assessed,
should be the same for each person. Eg., if \code{t = 0} (\code{t[1]})
corresponds to an index visit held by the variable \code{VisitDate}, then
\code{startPoint} should be set to \code{VisitDate}.}

\item{t}{numeric vector of time points to collect measurements, eg.
\code{c(0, 182.5, 365)} for measure collection at \code{t = 0},
\code{t = 180} (6 months from \code{t = 0}), and \code{t = 365} (1 year
from \code{t = 0}). Default is \code{c(0, 182.5, 365)} according to Janney
et al. 2016}

\item{windows}{numeric vector of measurement collection windows to use around
each time point in \code{t}. Eg. Janney et al. 2016 use
\code{c(30, 60, 60)} for \code{t} of \code{c(0, 182.5, 365)}, implying that
the closest measurement \code{t = 0} will be collected 30 days prior to and
30 days post \code{startPoint}. Subsequent measurements will be collected
60 days prior to and 60 days post t0+182.5 days, and t0+365 days.}
}
\value{
returns a data frame reduced by the inclusion criteria defined in
  \code{t} and \code{windows}, the resultant dimensions beng
  \code{ncol(df) + 2} columns and up to \code{length(t)} rows per group.
}
\description{
\code{meas_collect} For grouped time series, (e.g., per person) define time
points and collect measurements, optionally applying windows for possible
data points for each time point.
}
