% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noel2012.R
\name{noel}
\alias{noel}
\title{Noel et al. 2012 Measurment Cleaning Algorithm}
\usage{
noel(
  df,
  id,
  measures,
  tmeasures,
  outliers = c(70, 700),
  fiscal_start = 10,
  collapse = FALSE
)
}
\arguments{
\item{df}{object of class \code{data.frame}, containing \code{id} and
\code{measures}.}

\item{id}{string corresponding to the name of the column of group
identifiers in \code{df}.}

\item{measures}{string corresponding to the name of the column of
measurements in \code{df}.}

\item{tmeasures}{string corresponding to the name of the column of
measurement collection dates or times in \code{df}. If \code{tmeasures} is
a \code{Date} object, there may be more than one measurement on the same
day, if it is a precise date-time object (\code{?POSIXlt}), there may not
be more than one measurement on the same day.}

\item{outliers}{object of type \code{list} with numeric inputs corresponding
to the upper and lower bound for each time entry. Default is
\code{list(LB = 70, UB = 700)}.}

\item{fiscal_start}{integer to be passed to \code{lubridate::quarter()}.
Defaults to 10, indicating October, the Federal Fiscal Year starting month.}

\item{collapse}{= FALSE - aggregate to Fiscal Quarter? Default is
\code{FALSE}, returning \code{df} appended with \code{Qmedian} to be
collapsed after calling the function.}
}
\value{
data frame with the same \code{nrow(df)} number of rows, but
  \code{ncol(df) + 3} columns. \code{output}, gives the \code{measures} with
  outliers flagged as \code{NA}; \code{FYQ}, numeric fiscal-quarter
  (\code{dbl}) in YYYY.Q format, e.g., 1st quarter of 2016 is 2016.1;
  \code{Qmedian}, median computed for each \code{id} and \code{FYQ}.
}
\description{
For grouped time series, (e.g., per person), \code{noel} collects all
measurements within a certain time frame, then categorizes each time frame
into Fiscal Quarters. Then, for each person and each Fiscal Quarter,
calculate the median measurement over all measurements by Fiscal Quarter.
}
\examples{
library(dplyr)
library(ggplot2)
data(cdw1000)

noel_df <- noel(
   df = cdw1000,
   id = "id",
   measures = "Weight",
   tmeasures = "WeightDate",
   collapse = TRUE
  )

# dplyr::glimpse(noel_df)

samp <- noel_df \%>\%
  distinct(id) \%>\%
  sample_n(16) \%>\%
  pull()

noel_df \%>\%
  filter(id \%in\% samp) \%>\%
  ggplot(aes(x = FYQ, y = Qmedian)) +
  geom_line(color = "black") +
  geom_point(color = "green") +
  facet_wrap(vars(id), ncol = 4, scales = "free_y")
}
