% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maciejewski2016.R
\name{prep_maciejewski}
\alias{prep_maciejewski}
\title{Maciejewski Algorithm Part I.a. - Check the data frame and parameters for
conformity with expected structure}
\usage{
prep_maciejewski(df, id, measures, tmeasures, varvec = NULL)
}
\arguments{
\item{df}{object of class \code{data.frame}, containing \code{id} and
\code{measures}}

\item{id}{string corresponding to the name of the column of patient
identifiers in \code{df}.}

\item{measures}{string corresponding to the name of the column of measures in
\code{df}, e.g., numeric weight data if using to clean weight data.}

\item{tmeasures}{string corresponding to the name of the column of measure
dates and/or times in \code{df}}

\item{varvec}{optional. String vector containing \code{df}, \code{id},
\code{measures}, and \code{tmeasures}.}
}
\value{
\code{data.table} with properly formatted columns for downstream
  processing/weight cleaning.
}
\description{
\code{prep_maciejewski} If \code{varvec} was provided and none of \code{id},
\code{tmeasures}, and \code{measures} was provided, confirm suitability of
\code{varvec} argument value for use in this function. If \code{varvec} is
not specified, confirm user has provided variable names. Then confirm
specified variable names exist in input \code{df}. Creates a working copy of
input data, if \code{df} is not already a \code{data.table}, then coerce to
\code{data.table} and subset to just the 3 columns needed. Function then
confirms variables are not unusable data types and returns \code{data.table}.
}
\examples{
\dontrun{
x <- 1:5

# check each for errors
prep_maciejewski(df = x, varvec = c('a', 's', 'd'))
prep_maciejewski(df = iris, varvec = c('a', 's', 'd'))
prep_maciejewski(df = iris, varvec = c('Sepal.Length', 's', 'd'))
prep_maciejewski(df = iris, varvec = c('Sepal.Length', 'Sepal.Width', 'd'))
prep_maciejewski(
   df = iris,
   varvec = c('Sepal.Length', 'Sepal.Width', 'Petal.Width')
) # success

air.df <- airquality
air.df$date <- with(air.df, as.Date(paste('2020', Month, Day, sep = '-')))
air.df$charval <- as.character(air.df$Ozone)
prep_maciejewski(df = air.df, varvec = c('a', 's', 'd'))
prep_maciejewski(df = air.df, varvec = c('Temp', 's', 'd'))
prep_maciejewski(df = air.df, varvec = c('Temp', 'date', 'd'))
prep_maciejewski(df = air.df, varvec = c('Temp', 'date', 'charval'))
prep_maciejewski(df = air.df, varvec = c('Temp', 'date', 'Wind')) # success
}
}
